bash

+++ exp/flush_reload/run_hpc_fr_gpg.sh
@@ -1,11 +1,14 @@
 #!/bin/bash
 
+source ~/.bashrc
+
 ROOT_DIR="/home/zechengh/Mastik"
 EXP_ROOT_DIR=$ROOT_DIR/exp
 source $EXP_ROOT_DIR/exp_funcs.sh
 
-OUTPUT_FOLDER="./results"
-rm -r $OUTPUT_FOLDER
+OUTPUT_FOLDER=EXP_ROOT_DIR/flush_reload/results
+
+[ -d $OUTPUT_FOLDER ] && rm -r $OUTPUT_FOLDER
 mkdir -p $OUTPUT_FOLDER
 
 GPG=$ROOT_DIR/gnupg-1.4.13/g10/gpg
